requirements.txt

streamlit
opencv-python-headless
Pillow
numpy
